fix: correctly update time survey was taken (#8276)
* Update useSurveyNotification.ts

* move key name into constant

* Update constants.tsx

* rename constant

* remove unneeded Recoil stuff

* Update application.ts

* Update useSurveyNotification.ts
