+++ Composer/packages/client/src/components/Notifications/useSurveyNotification.ts
@@ -7,9 +7,9 @@ import formatMessage from 'format-message';
 import querystring from 'query-string';
 
 import { ClientStorage } from '../../utils/storage';
-import { surveyEligibilityState, dispatcherState, machineInfoState } from '../../recoilModel/atoms/appState';
+import { dispatcherState, machineInfoState } from '../../recoilModel/atoms/appState';
 import { MachineInfo } from '../../recoilModel/types';
-import { SURVEY_URL_BASE } from '../../constants';
+import { LAST_SURVEY_KEY, SURVEY_URL_BASE, SURVEY_PARAMETERS } from '../../constants';
 import TelemetryClient from '../../telemetry/TelemetryClient';
 
 const buildUrl = (info: MachineInfo) => {
@@ -30,16 +30,46 @@ const buildUrl = (info: MachineInfo) => {
   return `${SURVEY_URL_BASE}?${querystring.stringify(parameters)}`;
 };
 
+const getSurveyEligibility = () => {
+  const surveyStorage = new ClientStorage(window.localStorage, 'survey');
+
+  const optedOut = surveyStorage.get('optedOut', false);
+  if (optedOut) {
+    return false;
+  }
+
+  let days = surveyStorage.get('days', 0);
+  const lastUsed = surveyStorage.get('dateLastUsed', null);
+  const lastTaken = surveyStorage.get(LAST_SURVEY_KEY, null);
+  const today = new Date().toDateString();
+  if (lastUsed !== today) {
+    days += 1;
+    surveyStorage.set('days', days);
+  }
+  surveyStorage.set('dateLastUsed', today);
+
+  if (
+    // To be eligible for the survey, the user needs to have used Composer
+    // some minimum number of days.
+    days >= SURVEY_PARAMETERS.daysUntilEligible &&
+    // Also, either the user must have never taken the survey before or
+    // the last time they took it must be long enough in the past.
+    (lastTaken == null || Date.now() - lastTaken > SURVEY_PARAMETERS.timeUntilNextSurvey)
+  ) {
+    // If the above conditions are true, there's a fixed chance the card will appear.
+    return Math.random() < SURVEY_PARAMETERS.chanceToAppear;
+  }
+};
+
 export const useSurveyNotification = () => {
   const { addNotification, deleteNotification } = useRecoilValue(dispatcherState);
-  const surveyEligible = useRecoilValue(surveyEligibilityState);
   const machineInfo = useRecoilValue(machineInfoState);
 
   useEffect(() => {
     const url = buildUrl(machineInfo);
     deleteNotification('survey');
 
-    if (surveyEligible) {
+    if (getSurveyEligibility()) {
       const surveyStorage = new ClientStorage(window.localStorage, 'survey');
       TelemetryClient.track('HATSSurveyOffered');
 
@@ -55,6 +85,7 @@ export const useSurveyNotification = () => {
               // This is safe; we control the URL that gets built
               // eslint-disable-next-line security/detect-non-literal-fs-filename
               window.open(url, '_blank');
+              surveyStorage.set(LAST_SURVEY_KEY, Date.now());
               TelemetryClient.track('HATSSurveyAccepted');
               deleteNotification('survey');
             },
