+++ Composer/packages/client/src/recoilModel/atoms/appState.ts
@@ -366,11 +366,6 @@ export const warnAboutFunctionsState = atom<boolean>({
   default: false,
 });
 
-export const surveyEligibilityState = atom<boolean>({
-  key: getFullyQualifiedKey('surveyEligibilityState'),
-  default: false,
-});
-
 export const machineInfoState = atom<MachineInfo>({
   key: getFullyQualifiedKey('machineInfoState'),
   default: null,
